Initial commit for parent User class
