+++ app/src/main/java/com/example/impact/model/Admin.java
@@ -1,5 +1,13 @@
 package com.example.impact.model;
 
-public class Admin {
+import androidx.annotation.Nullable;
+
+public class Admin extends User {
+    public Admin() {
+        // Default constructor for Firestore
+    }
+    public Admin(String id, @Nullable String name, String email, @Nullable String phone) {
+        super(id, name, email, phone);
+    }
 
 }
