+++ app/src/main/java/com/example/impact/model/Organizer.java
@@ -1,4 +1,13 @@
 package com.example.impact.model;
 
-public class Organizer {
+import androidx.annotation.Nullable;
+
+public class Organizer extends User {
+    public Organizer() {
+        // Default constructor for Firestore
+    }
+
+    public Organizer(String id, @Nullable String name, String email, @Nullable String phone) {
+        super(id, name, email, phone);
+    }
 }
